requirements.txt
# wort to excel
# Usage: pip install -r requirements.txt

# Base ------------------------------------------------------------------------

python-docx>=0.8.11
streamlit>=1.24.0
pandas>=2.0.3
XlsxWriter>=3.1.2